+++ src/array/index.ts
@@ -1,0 +1,3 @@
+export * from './removeEmptyArray';
+export * from './uniqueArray';
+export * from './uniqueArrayObject';