+++ src/array/uniqueArray.ts
@@ -1,0 +1,3 @@
+export const uniqueArray = (array: any[]) => {
+  return [...new Set(array)];
+};