+++ src/string/stripcslashes.ts
@@ -1,0 +1,6 @@
+export const stripcslashes = (text: string): any => {
+  const renderedString = text.replace(/\\./g, (match) => {
+    return new Function('return "' + match + '"')() || match;
+  });
+  return renderedString;
+};