+++ src/url/isValidHttpUrl.ts
@@ -1,0 +1,10 @@
+export function isValidHttpUrl(value: string): boolean {
+  let url;
+  try {
+    url = new URL(value);
+  } catch (_) {
+    return false;
+  }
+
+  return url.protocol === 'http:' || url.protocol === 'https:';
+}