+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## v1.0.2
+
+### Added or Changed
+
+- Add removeNullObject
+
+### Removed
+
+## v1.0.1
+
+### Added or Changed
+
+- Add Validate
+- Add Array Utils
+- Add Object Utils
+- Add String Utils
+- Add Number Utils
+- Add Ip Utils
+- Add Url Utils
+
+### Removed
+
 ## v1.0.1
 
 ### Added or Changed